apps/projects: Move project deleted notification to notifications
There doesn't seem to be a obvious resean why this is handled separately.
