+++ apps/notifications/emails.py
@@ -1,5 +1,7 @@
 from django.contrib import auth
 
+from apps.organisations.models import Organisation
+from apps.projects import tasks
 from apps.users.emails import EmailAplus as Email
 
 User = auth.get_user_model()
@@ -103,6 +105,37 @@ class NotifyInitiatorsOnProjectCreatedEmail(Email):
         creator = User.objects.get(pk=self.kwargs['creator_pk'])
         context['creator'] = creator
         context['project'] = self.object
+        return context
+
+
+class NotifyInitiatorsOnProjectDeletedEmail(Email):
+    template_name = \
+        'a4_candy_notifications/emails/notify_initiators_project_deleted'
+
+    @classmethod
+    def send_no_object(cls, object, *args, **kwargs):
+        organisation = object.organisation
+        object_dict = {
+            'name': object.name,
+            'initiators': list(organisation.initiators.all()
+                               .distinct()
+                               .values_list('email', flat=True)),
+            'organisation_id': organisation.id
+        }
+        tasks.send_async_no_object(
+            cls.__module__, cls.__name__,
+            object_dict, args, kwargs)
+        return []
+
+    def get_organisation(self):
+        return Organisation.objects.get(id=self.object['organisation_id'])
+
+    def get_receivers(self):
+        return self.object['initiators']
+
+    def get_context(self):
+        context = super().get_context()
+        context['name'] = self.object['name']
         return context
