+++ apps/notifications/templates/a4_candy_notifications/emails/notify_initiators_project_deleted.en.email
